requirements.txt
Flask==3.1.1
SQLAlchemy==2.0.43
python-dotenv==1.1.1
gunicorn==21.2.0